Move mapping extension classes
