+++ src/src.sln
@@ -1,0 +1,6 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.0.0
+Global
+EndGlobal
